requirements.txt
dicord
discord[voice]
pillow
